misc changes
- merge `ui` example with `styled-map`
- don't use parameters that are null for animateCamera and moveCamera
- add removePinchOnPressed to MapCompass
- refactor
