+++ lib/src/map.dart
@@ -27,6 +27,12 @@ class MapLibreMap extends StatefulWidget {
 
   /// Flutter widgets that get displayed on top on the map and are within the
   /// [MapLibreMap] context.
+  ///
+  /// You can use the following included UI elements:
+  /// - [MapCompass]
+  /// - [MapControlButtons]
+  /// - [MapScalebar]
+  /// - [SourceAttribution].
   final List<Widget> children;
 
   /// Which gestures should be consumed by the map.
